requirements.txt
Flask==3.0.0
flask-cors==4.0.0
transformers==4.36.2
torch==2.1.2
tokenizers==0.15.0
sentence-transformers==2.3.1
huggingface_hub==0.19.4
nltk==3.8.1
spacy==3.7.2
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.2
scipy==1.11.4
textblob==0.17.1
langdetect==1.0.9
matplotlib==3.7.5
plotly==5.17.0
seaborn==0.13.0
networkx==3.2.1
matplotlib-venn==0.11.10
requests==2.31.0
Pillow==10.1.0
faiss-cpu==1.7.4
wordcloud==1.9.3
psutil==5.9.6
